Initial commit
Just testing out the API so far, seeing what it can accomplish. Quite a bit really, very impressive stuff.

I got a little lost in the SCSS so I made a to do list for myself. A sort of development road map for this project. That should help keep me on track.

Excited for this one :)